pyproject.toml
[tool.poetry]
name = "polls"
version = "0.1"
description = "Voting microservice"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
orjson = "^3.10.15"
pydantic = "^2.10.5"
python-dotenv = "^1.0.1"
SQLAlchemy = "^2.0.37"
uvicorn = "^0.34.0"
alembic = "^1.14.0"
pytest = "^8.3.4"
loguru = "^0.7.3"
tenacity = "^9.0.0"
psycopg = "^3.2.4"
typing-extensions = "^4.12.2"
PyJWT = "^2.11.0"
cryptography = "^46.0.0"
httpx = "^0.28.0"
apscheduler = "^3.11.0"
asyncpg = "^0.30.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
mypy = "^1.14.1"

